Port of b2883ec
https://github.com/gama-platform/gama/commit/b2883eca1c023fa818fb99781b9305c32bb47b2b

+++ gama.kernel/gama.core/src/gama/core/kernel/batch/exploration/morris/MorrisExploration.java
@@ -1,7 +1,7 @@
 /*******************************************************************************************************
  *
- * MorrisExploration.java, in msi.gama.core, is part of the source code of the GAMA modeling and simulation platform
- * (v.1.9.0).
+ * MorrisExploration.java, in gama.core, is part of the source code of the GAMA modeling and simulation platform
+ * (v.1.9.2).
  *
  * (c) 2007-2023 UMI 209 UMMISCO IRD/SU & Partners (IRIT, MIAT, TLU, CTU)
  *
@@ -21,8 +21,6 @@ import java.util.List;
 import java.util.Map;
 
 import gama.annotations.common.interfaces.IKeyword;
-import gama.annotations.precompiler.IConcept;
-import gama.annotations.precompiler.ISymbolKind;
 import gama.annotations.precompiler.GamlAnnotations.doc;
 import gama.annotations.precompiler.GamlAnnotations.example;
 import gama.annotations.precompiler.GamlAnnotations.facet;
@@ -30,11 +28,15 @@ import gama.annotations.precompiler.GamlAnnotations.facets;
 import gama.annotations.precompiler.GamlAnnotations.inside;
 import gama.annotations.precompiler.GamlAnnotations.symbol;
 import gama.annotations.precompiler.GamlAnnotations.usage;
+import gama.annotations.precompiler.IConcept;
+import gama.annotations.precompiler.ISymbolKind;
 import gama.core.common.util.FileUtils;
 import gama.core.kernel.batch.exploration.AExplorationAlgorithm;
 import gama.core.kernel.batch.exploration.sampling.MorrisSampling;
-import gama.core.kernel.experiment.ParametersSet;
+import gama.core.kernel.experiment.BatchAgent;
 import gama.core.kernel.experiment.IParameter.Batch;
+import gama.core.kernel.experiment.ParameterAdapter;
+import gama.core.kernel.experiment.ParametersSet;
 import gama.core.runtime.IScope;
 import gama.core.runtime.concurrent.GamaExecutorService;
 import gama.core.runtime.exceptions.GamaRuntimeException;
@@ -174,7 +176,8 @@ public class MorrisExploration extends AExplorationAlgorithm {
 		}
 		/* Disable repetitions / repeat argument */
 		currentExperiment.setSeeds(new Double[1]);
-		// TODO : why doesnt it take into account the value of 'keep_simulations:' ? because, by design, there is to many simulation to keep in memory... 
+		// TODO : why doesnt it take into account the value of 'keep_simulations:' ? because, by design, there is to
+		// many simulation to keep in memory...
 		currentExperiment.setKeepSimulations(false);
 		if (GamaExecutorService.shouldRunAllSimulationsInParallel(currentExperiment)) {
 			res_outputs = currentExperiment.launchSimulationsWithSolution(solutions);
@@ -184,31 +187,36 @@ public class MorrisExploration extends AExplorationAlgorithm {
 				res_outputs.put(sol, currentExperiment.launchSimulationsWithSolution(sol));
 			}
 		}
-		
+
 		// The output of simulations
 		Map<String, List<Double>> rebuilt_output = rebuildOutput(scope, res_outputs);
-		List<String> output_names = rebuilt_output.keySet().stream().toList();
-		
-		String path_to = Cast.asString(scope, getFacet(IKeyword.BATCH_REPORT).value(scope));
-		final File fm = new File(FileUtils.constructAbsoluteFilePath(scope, path_to, false));
-		final File parentm = fm.getParentFile();
-		if (!parentm.exists()) { parentm.mkdirs(); }
-		if (fm.exists()) { fm.delete(); }
-		
-		for (int i = 0; i < rebuilt_output.size(); i++) {
-			String tmp_name = output_names.get(i);
-			List<Map<String, Double>> morris_coefficient =
-					Morris.MorrisAggregation(nb_levels, rebuilt_output.get(tmp_name), MySamples);
-			Morris.WriteAndTellResult(tmp_name, fm.getAbsolutePath(), scope, morris_coefficient);
-		}
-		/* Save the simulation values in the provided .csv file (input and corresponding output) */
-		if (hasFacet(IKeyword.BATCH_OUTPUT)) {
-			path_to = Cast.asString(scope, getFacet(IKeyword.BATCH_OUTPUT).value(scope));
-			final File fo = new File(FileUtils.constructAbsoluteFilePath(scope, path_to, false));
-			final File parento = fo.getParentFile();
-			if (!parento.exists()) { parento.mkdirs(); }
-			if (fo.exists()) { fo.delete(); }
-			saveSimulation(rebuilt_output, fo, scope);
+
+		// Prevent OutOfBounds when experiment ends before morris exploration is completed
+		if (res_outputs.size() == MySamples.size()) {
+
+			List<String> output_names = rebuilt_output.keySet().stream().toList();
+
+			String path_to = Cast.asString(scope, getFacet(IKeyword.BATCH_REPORT).value(scope));
+			final File fm = new File(FileUtils.constructAbsoluteFilePath(scope, path_to, false));
+			final File parentm = fm.getParentFile();
+			if (!parentm.exists()) { parentm.mkdirs(); }
+			if (fm.exists()) { fm.delete(); }
+
+			for (int i = 0; i < rebuilt_output.size(); i++) {
+				String tmp_name = output_names.get(i);
+				List<Map<String, Double>> morris_coefficient =
+						Morris.MorrisAggregation(nb_levels, rebuilt_output.get(tmp_name), MySamples);
+				Morris.WriteAndTellResult(tmp_name, fm.getAbsolutePath(), scope, morris_coefficient);
+			}
+			/* Save the simulation values in the provided .csv file (input and corresponding output) */
+			if (hasFacet(IKeyword.BATCH_OUTPUT)) {
+				path_to = Cast.asString(scope, getFacet(IKeyword.BATCH_OUTPUT).value(scope));
+				final File fo = new File(FileUtils.constructAbsoluteFilePath(scope, path_to, false));
+				final File parento = fo.getParentFile();
+				if (!parento.exists()) { parento.mkdirs(); }
+				if (fo.exists()) { fo.delete(); }
+				saveSimulation(rebuilt_output, fo, scope);
+			}
 		}
 	}
 
@@ -228,7 +236,30 @@ public class MorrisExploration extends AExplorationAlgorithm {
 		this.MySamples = Cast.asList(scope, morris_samplings.get(0));
 		return Cast.asList(scope, morris_samplings.get(1));
 	}
-	
+
+	@Override
+	public void addParametersTo(final List<Batch> exp, final BatchAgent agent) {
+		super.addParametersTo(exp, agent);
+
+		int s = Cast.asInt(agent.getScope(), getFacet(SAMPLE_SIZE).value(agent.getScope()));
+		int l = Cast.asInt(agent.getScope(), getFacet(NB_LEVELS).value(agent.getScope()));
+
+		exp.add(new ParameterAdapter("Morris level", IKeyword.MORRIS, IType.STRING) {
+			@Override
+			public Object value() {
+				return l;
+			}
+		});
+
+		exp.add(new ParameterAdapter("Morris sample", IKeyword.MORRIS, IType.STRING) {
+			@Override
+			public Object value() {
+				return solutions == null ? s * l : solutions.size() == 0 ? sample : solutions.size();
+			}
+		});
+
+	}
+
 	// ************************* END OF SUPER CLASS CONTRACT ************************* //
 
 	/**
@@ -246,7 +277,11 @@ public class MorrisExploration extends AExplorationAlgorithm {
 		for (String output : outputs) { rebuilt_output.put(output, new ArrayList<>()); }
 		for (ParametersSet sol : solutions) {
 			for (String output : outputs) {
-				rebuilt_output.get(output).add(Cast.asFloat(scope, res_outputs.get(sol).get(output).get(0)));
+				try {
+					rebuilt_output.get(output).add(Cast.asFloat(scope, res_outputs.get(sol).get(output).get(0)));
+				} catch (NullPointerException e) {
+					return rebuilt_output;
+				}
 			}
 		}
 		return rebuilt_output;
